+++ cdh-tsat6-stm32project-nucleo/Drivers/Hardware_Peripherals/Inc/AS3001204_driver.h
@@ -22,56 +22,56 @@
 //  Define Directives
 // ###############################################################################################
 
-#define AS3001204_SPI hspi2
-#define AS3001204_SPI_DELAY HAL_MAX_DELAY
+#define AS3001204_SPI 			hspi2
+#define AS3001204_SPI_DELAY 	HAL_MAX_DELAY
 
-#define AS3001204_nCS_GPIO GPIOC
-#define AS3001204_nCS_PIN GPIO_PIN_8
+#define AS3001204_nCS_GPIO 		GPIOC
+#define AS3001204_nCS_PIN 		GPIO_PIN_8
 
-#define AS3001204_nWP_GPIO GPIOC
-#define AS3001204_nWP_PIN GPIO_PIN_9
+#define AS3001204_nWP_GPIO 		GPIOC
+#define AS3001204_nWP_PIN 		GPIO_PIN_9
 
 // Opcodes (datasheet pp. 32-36)
 // Control operations (1-0-0 type)
-#define AS3001204_OPCODE_WRITE_ENABLE 0x06
-#define AS3001204_OPCODE_WRITE_DISABLE 0x04
-#define AS3001204_OPCODE_ENTER_DEEP_PWDOWN 0xb9
-#define AS3001204_OPCODE_ENTER_HIBERNATE 0xba
-#define AS3001204_OPCODE_EXIT_DEEP_PWDOWN 0xab
-#define AS3001204_OPCODE_SOFT_RESET_ENABLE 0x66
-#define AS3001204_OPCODE_SOFT_RESET 0x99
+#define AS3001204_OPCODE_WRITE_ENABLE		0x06
+#define AS3001204_OPCODE_WRITE_DISABLE		0x04
+#define AS3001204_OPCODE_ENTER_DEEP_PWDOWN	0xb9
+#define AS3001204_OPCODE_ENTER_HIBERNATE	0xba
+#define AS3001204_OPCODE_EXIT_DEEP_PWDOWN	0xab
+#define AS3001204_OPCODE_SOFT_RESET_ENABLE	0x66
+#define AS3001204_OPCODE_SOFT_RESET			0x99
 
 // Read register operations (1-0-1 type)
-#define AS3001204_OPCODE_READ_STATUS_REG 0x05
-#define AS3001204_OPCODE_READ_CONFIG_REGS 0x46
-#define AS3001204_OPCODE_READ_DEVICE_ID 0x9f
-#define AS3001204_OPCODE_READ_UNIQUE_ID 0x4c
-#define AS3001204_OPCODE_READ_AAP_REG 0x14
+#define AS3001204_OPCODE_READ_STATUS_REG 	0x05
+#define AS3001204_OPCODE_READ_CONFIG_REGS 	0x46
+#define AS3001204_OPCODE_READ_DEVICE_ID 	0x9f
+#define AS3001204_OPCODE_READ_UNIQUE_ID 	0x4c
+#define AS3001204_OPCODE_READ_AAP_REG 		0x14
 
 // Write register operations (1-0-1 type)
-#define AS3001204_OPCODE_WRITE_STATUS_REG 0x01
-#define AS3001204_OPCODE_WRITE_CONFIG_REGS 0x87
-#define AS3001204_OPCODE_WRITE_AAP_REG 0x1a
+#define AS3001204_OPCODE_WRITE_STATUS_REG 	0x01
+#define AS3001204_OPCODE_WRITE_CONFIG_REGS 	0x87
+#define AS3001204_OPCODE_WRITE_AAP_REG 		0x1a
 
 // Register lengths (in bytes)
-#define AS3001204_STATUS_REG_LENGTH 1
-#define AS3001204_CONFIG_REGS_LENGTH 4
-#define AS3001204_DEVICE_ID_LENGTH 4
-#define AS3001204_UNIQUE_ID_LENGTH 8
-#define AS3001204_AAP_REG_LENGTH 1
+#define AS3001204_STATUS_REG_LENGTH 		1
+#define AS3001204_CONFIG_REGS_LENGTH 		4
+#define AS3001204_DEVICE_ID_LENGTH 			4
+#define AS3001204_UNIQUE_ID_LENGTH 			8
+#define AS3001204_AAP_REG_LENGTH 			1
 
 // Memory operations (1-1-1 type)
-#define AS3001204_OPCODE_READ_MEMORY 0x03
-#define AS3001204_OPCODE_WRITE_MEMORY 0x02
-#define AS3001204_OPCODE_READ_AUG_STORAGE 0x4b
-#define AS3001204_OPCODE_WRITE_AUG_STORAGE 0x42
+#define AS3001204_OPCODE_READ_MEMORY 		0x03
+#define AS3001204_OPCODE_WRITE_MEMORY 		0x02
+#define AS3001204_OPCODE_READ_AUG_STORAGE 	0x4b
+#define AS3001204_OPCODE_WRITE_AUG_STORAGE 	0x42
 
 // Delay bytes to await response from augmented storage array
 // (see timing diagram, datasheet pp. 38-39)
-#define AS3001204_READ_AUG_STORAGE_DELAY {0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00};
+#define AS3001204_READ_AUG_STORAGE_DELAY 	{0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00};
 
 // Device ID Values
-#define AS3001204_DEVICE_ID 0xE6010102
+#define AS3001204_DEVICE_ID 				0xE6010102
 
 // ###############################################################################################
 //  Global Variable Declarations
