+++ cdh-tsat6-stm32project-nucleo/Drivers/Hardware_Peripherals/Test/Src/AS3001204_driver_test.c
@@ -13,7 +13,7 @@
 
 
 #include "../Inc/AS3001204_driver.h"
-#include "AS3001204_driver_test.h"
+#include "../Inc/AS3001204_driver_test.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
